done: DOM-2340 Modified Selection
